app/pyproject.toml
[project]
name = "flask-app"
version = "0.1.0"
dependencies = [
    "flask>=3.0.3",
    "keras>=3.6.0",
    "numpy>=2.0.2",
    "pillow>=11.0.0",
    "scikit-learn>=1.6.0",
    "tensorflow>=2.18.0",
    "mlflow>=2.19.0",
    "prometheus-client>=0.21.1",
    "prometheus-flask-exporter>=0.23.1",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[tool.setuptools]
py-modules = []